requirements.txt
invoke
tornado
markdown
pymdown-extensions
python-dotenv
aiobotocore
boto3
tld
cryptography
qrcode
liteblue